+++ backend/routes.js
@@ -4,6 +4,10 @@ const router = express.Router();
 const mongoose = require("mongoose");
 
 
+
+
+
+
 router.get("/", async (req, res) => {
     try {
         const schools = await AnimalSchool.find();
