+++ frontend/src/About.jsx
@@ -1,0 +1,21 @@
+import React from 'react'
+
+const About = () => {
+  return (
+    <div>
+        <h1>Animal Schools</h1>
+      
+     <div style={{ fontFamily: "Arial, sans-serif", padding: "20px" }}>
+      <h2 style={{ textAlign: "center", color: "#4CAF50" }}>Animal Schools</h2>
+      <p style={{ textAlign: "center", color: "#333" }}>
+        Animal schools are unique learning institutions where different species 
+        of animals develop survival skills, communication techniques, and social behaviors.
+        These schools exist in diverse environments such as jungles, oceans, savannahs, 
+        and arctic regions, each adapted to the specific needs of its inhabitants.
+      </p>
+    </div>
+    </div>
+  )
+}
+
+export default About
